% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_mod.R
\name{ca_server}
\alias{ca_server}
\title{Server backend for CA modules}
\usage{
ca_server(id, data, input, output)
}
\arguments{
\item{id}{x}

\item{data}{x}

\item{input}{x}

\item{output}{x}
}
\value{
xx
}
\description{
Server backend for CA modules
}
