% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evp_mod.R
\name{evp_server}
\alias{evp_server}
\title{Server backend for EVP module}
\usage{
evp_server(id, data, input, output)
}
\arguments{
\item{id}{x}

\item{data}{x}

\item{input}{x}

\item{output}{x}
}
\value{
xx
}
\description{
Server backend for EVP module
}
