% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_squba_tbls.R
\name{format_squba_tbls}
\alias{format_squba_tbls}
\title{Format SQUBA tables}
\usage{
format_squba_tbls(squba_rslt_directory = Sys.getenv("squba_rslt_dir__"))
}
\arguments{
\item{squba_rslt_directory}{result directory where SQUBA result CSVs live}
}
\value{
list of dataframes that will be used by the app
}
\description{
Format SQUBA tables
}
