% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funcs.R
\name{load_check_data}
\alias{load_check_data}
\title{Load data + filter data for processing}
\usage{
load_check_data(dat_list, input, type_input)
}
\arguments{
\item{dat_list}{the list with relevant dataframes for the module}

\item{input}{app-level input}

\item{type_input}{the name of the check type input field for the module}
}
\value{
the correct dataframe based on the selected check type for the module of interest
}
\description{
Load data + filter data for processing
}
