% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_mod.R
\name{pf_server}
\alias{pf_server}
\title{Server backend for PF module}
\usage{
pf_server(id, data, input, output)
}
\arguments{
\item{id}{x}

\item{data}{x}

\item{input}{x}

\item{output}{x}
}
\value{
XX
}
\description{
Server backend for PF module
}
