% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funcs.R
\name{plot_downloader}
\alias{plot_downloader}
\title{Plot downloader}
\usage{
plot_downloader(plt_obj, id)
}
\arguments{
\item{plt_obj}{a reactive list of all plots for the module + check}

\item{id}{the ID of the module}
}
\value{
a zip file will all plots displayed in the UI
}
\description{
Plot downloader
}
