% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funcs.R
\name{plt_engine_selector}
\alias{plt_engine_selector}
\title{Select appropriate engine & render plot}
\usage{
plt_engine_selector(plt_rctv, n_plt)
}
\arguments{
\item{plt_rctv}{the reactive plot object generated by a SQUBA *_output function}

\item{n_plt}{the N of the plot being generated server side (1, 2, or 3)}
}
\value{
the rendered plot based on the procedurally selected engine
}
\description{
Select appropriate engine & render plot
}
