% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funcs.R
\name{render_update_notes}
\alias{render_update_notes}
\title{Render & update notes table}
\usage{
render_update_notes(notes_rctval, id, mod_name, other_inps, input)
}
\arguments{
\item{notes_rctval}{reactive tibble with notes information}

\item{id}{the ID of the module for which notes are being taken}

\item{mod_name}{the full name of the module}

\item{other_inps}{any module-specific inputs that need to be documented}

\item{input}{the server-level input}
}
\value{
a reactive table with all notes information from the session
}
\description{
Render & update notes table
}
