% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_render.R
\name{squba_dive}
\alias{squba_dive}
\title{SQUBA Dive}
\usage{
squba_dive(squba_results_directory = "../dummy_rslts")
}
\arguments{
\item{squba_results_directory}{the local directory where results in CSV format are stored}
}
\value{
a Shiny app displaying all of (and limited to only) the results in
the input directory
}
\description{
Render the Shiny interface for SQUBA Diver. This function will
read all CSV files in the indicated directory and render a bespoke
interface displaying the results. Users can then use the reactively
generated toggles to explore the data in different ways, like viewing
different numeric columns on the plot or utilizing the large N
functionality. Users can also take notes on the plot and save them to
a reactive table, or download the plots to save and use in manuscripts.
}
\examples{
\dontrun{
squba_dive('path/to/results')
}

`Load data`
}
