% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funcs.R
\name{update_type_input}
\alias{update_type_input}
\title{Update Check Type selector drop down}
\usage{
update_type_input(dat_list, ck_str)
}
\arguments{
\item{dat_list}{the list with relevant dataframes for the module}

\item{ck_str}{the abbreviated string of the module to be used to reference the input field}
}
\value{
will update the existing check type input field for the module of interest
with the available check types based on the data
}
\description{
Update Check Type selector drop down
}
